updated unit tests adding conftest and moving fixtures to there for reuse
